+++ samples/dotnet/kernel-syntax-examples/Example17_ChatGPT.cs
@@ -16,9 +16,6 @@ public static class Example17_ChatGPT
 {
     public static async Task RunAsync()
     {
-        await AzureOpenAIChatStreamSampleAsync();
-        await OpenAIChatStreamSampleAsync();
-
         await AzureOpenAIChatSampleAsync();
         await OpenAIChatSampleAsync();
 
@@ -82,60 +79,6 @@ public static class Example17_ChatGPT
         await StartChatAsync(chatGPT);
     }
 
-    private static async Task OpenAIChatStreamSampleAsync()
-    {
-        Console.WriteLine("======== Open AI - ChatGPT Streaming ========");
-
-        IKernel kernel = new KernelBuilder().WithLogger(ConsoleLogger.Log).Build();
-
-        // Add your chat completion service
-        kernel.Config.AddOpenAIChatCompletionService("gpt-3.5-turbo", Env.Var("OPENAI_API_KEY"));
-
-        IChatCompletion chatGPT = kernel.GetService<IChatCompletion>();
-
-        await StartStreamingChatAsync(chatGPT);
-    }
-
-    private static async Task AzureOpenAIChatStreamSampleAsync()
-    {
-        Console.WriteLine("======== Azure Open AI - ChatGPT Streaming ========");
-
-        IKernel kernel = new KernelBuilder().WithLogger(ConsoleLogger.Log).Build();
-
-        // Add your chat completion service
-        kernel.Config.AddAzureChatCompletionService(
-            Env.Var("AZURE_OPENAI_CHAT_DEPLOYMENT_NAME"),
-            Env.Var("AZURE_OPENAI_ENDPOINT"),
-            Env.Var("AZURE_OPENAI_KEY"));
-
-        IChatCompletion chatGPT = kernel.GetService<IChatCompletion>();
-
-        await StartStreamingChatAsync(chatGPT);
-    }
-
-    private static async Task StartStreamingChatAsync(IChatCompletion chatGPT)
-    {
-        Console.WriteLine("Chat content:");
-        Console.WriteLine("------------------------");
-
-        var chatHistory = (OpenAIChatHistory)chatGPT.CreateNewChat("You are a librarian, expert about books");
-        await MessageOutputAsync(chatHistory);
-
-        // First user message
-        chatHistory.AddUserMessage("Hi, I'm looking for book suggestions");
-        await MessageOutputAsync(chatHistory);
-
-        // First bot assistant message
-        await StreamMessageOutputAsync(chatGPT, chatHistory);
-
-        // Second user message
-        chatHistory.AddUserMessage("I love history and philosophy, I'd like to learn something new about Greece, any suggestion?");
-        await MessageOutputAsync(chatHistory);
-
-        // Second bot assistant message
-        await StreamMessageOutputAsync(chatGPT, chatHistory);
-    }
-
     private static async Task StartChatAsync(IChatCompletion chatGPT)
     {
         Console.WriteLine("Chat content:");
@@ -173,21 +116,5 @@ public static class Example17_ChatGPT
         Console.WriteLine("------------------------");
 
         return Task.CompletedTask;
-    }
-
-    private static async Task StreamMessageOutputAsync(IChatCompletion chatGPT, ChatHistory chatHistory,
-        ChatHistory.AuthorRoles authorRole = ChatHistory.AuthorRoles.Assistant)
-    {
-        Console.Write($"{authorRole}: ");
-        string fullMessage = string.Empty;
-
-        await foreach (string message in chatGPT.GenerateMessageStreamAsync(chatHistory))
-        {
-            fullMessage += message;
-            Console.Write(message);
-        }
-
-        Console.WriteLine("\n------------------------");
-        chatHistory.AddMessage(authorRole, fullMessage);
     }
 }
