+++ samples/dotnet/kernel-syntax-examples/Program.cs
@@ -104,7 +104,10 @@ public static class Program
         await Example32_StreamingCompletion.RunAsync();
         Console.WriteLine("== DONE ==");
 
-        await Example33_CustomChat.RunAsync();
+        await Example33_StreamingChat.RunAsync();
+        Console.WriteLine("== DONE ==");
+
+        await Example34_CustomChatModel.RunAsync();
         Console.WriteLine("== DONE ==");
     }
 }
